+++ src/Elementos/Operacion.java
@@ -13,27 +13,27 @@ package Elementos;
  * @since 1.0
  */
 public abstract class Operacion {
-    private float operando1;
-    private float operando2;
+    private double operando1;
+    private double operando2;
 
-    public Operacion(float operando1, float operando2) {
+    public Operacion(double operando1, double operando2) {
         this.operando1 = operando1;
         this.operando2 = operando2;
     }
 
-    public float getOperando1() {
+    public double getOperando1() {
         return operando1;
     }
 
-    public void setOperando1(float operando1) {
+    public void setOperando1(double operando1) {
         this.operando1 = operando1;
     }
 
-    public float getOperando2() {
+    public double getOperando2() {
         return operando2;
     }
 
-    public void setOperando2(float operando2) {
+    public void setOperando2(double operando2) {
         this.operando2 = operando2;
     }
     
@@ -41,6 +41,6 @@ public abstract class Operacion {
      * Método para operar dos operandos, determinando su operador.
      * @return 
      */
-    public abstract float operar();
+    public abstract double operar();
  
 }
