+++ src/Elementos/Resta.java
@@ -14,12 +14,12 @@ package Elementos;
  */
 public class Resta extends Operacion{
 
-    public Resta(float operando1, float operando2) {
+    public Resta(double operando1, double operando2) {
         super(operando1, operando2);
     }
 
     @Override
-    public float operar() {
+    public double operar() {
         return  getOperando1() - getOperando2();
     }
